% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainBiocBuildFailure.R
\name{explainBiocBuildFailure}
\alias{explainBiocBuildFailure}
\title{Attempt to explain why a certain Bioconductor package fails on the build machine(s)}
\usage{
explainBiocBuildFailure(
  package,
  biocVersion = BiocManager::version(),
  buildNodes = c("malbec2", "tokay2", "riesling1", "machv2"),
  showOnlyFailing = FALSE,
  repo = "BioCsoft",
  dependencies = c("Depends", "Imports")
)
}
\arguments{
\item{package}{Character string giving the package of interest.}

\item{biocVersion}{Character string giving the Bioconductor version.}

\item{buildNodes}{Character vector indicating the build nodes to check.
Should be a subset of c("malbec2", "tokay2", "riesling1", "machv2").
If more than one node is specified, a package is considered to fail
if it fails on at least one of them. A failure is recognized by
an "ERROR" returned in the build report, in any of the stages.}

\item{showOnlyFailing}{Logical, whether to show all dependencies or
just the failing ones (and the intermediate nodes between the failing
ones and the package of interest) in the output graph.}

\item{repo}{Character vector indicating the repo(s) to consider. Will be
passed to \code{biocPkgList()}.}

\item{dependencies}{Character vector indicating the types of dependencies
to consider. Will be passed to \code{biocPkgList()}.}
}
\value{
Plots a graph containing the package of interest as well as its dependencies.
For each package, the graph displays the name, the date when it was
last changed, and the fraction of its direct descendants (packages that
depend on or import it) that currently fail. Note that only Bioc
dependencies are shown.
}
\description{
Attempt to explain why a certain Bioconductor package fails on the build machine(s)
}
\examples{
explainBiocBuildFailure("Rhisat2")
}
\author{
Charlotte Soneson
}
